feat: Update the number of movies in the "Mental Food" list
The "影剧" (Movies and TV Dramas) section of the "Mental Food" page has
been updated to show that there are now 39 movies listed, instead of 36.
This update expands the variety and selection of movies available for
mental entertainment. It ensures that users have more options to explore
and engage with. This change aims to enhance the user experience and
provide a richer source of mental stimulation.

Fixes #1234